requirements.light.txt
addict==2.4.0
basicsr==1.4.2
cityscapesscripts==2.2.4
einops==0.8.1
fairscale==0.4.13
fastapi==0.124.4
ftfy==6.3.1
fvcore==0.1.5.post20221221
gradio==6.1.0
h5py==3.15.1
httpx==0.28.1
hydra-core==1.3.2
imutils==0.5.4
iopath==0.1.10
lmdb==1.7.5
matplotlib==3.10.8
numpy>=1.23,<2.0
omegaconf==2.3.0
onnx==1.20.0
packaging==25.0
Pillow==12.0.0
prettytable==3.17.0
pycocotools==2.0.10
Pygments==2.19.2
python-dotenv==1.2.1
pytorch_lightning==2.6.0
PyYAML==6.0.3
regex==2024.11.6
requests==2.32.5
safetensors==0.7.0
scipy==1.13.1
Shapely==2.1.2
tabulate==0.9.0
tensorboardX==2.6.4
termcolor==3.2.0
tifffile==2024.8.30
tqdm==4.67.1
uvicorn==0.38.0
yapf==0.43.0
